Moved the data augmentations to the same file
